refcator evaluating all flags
